Code cleanup: reduce casting for appl_xxx() routines
The types of the parameters for routines called by gemsuper() for
appl_xxx() functions have been made natural; required casting is
now done in gemsuper.c.

+++ aes/gemwmlib.c
@@ -622,7 +622,7 @@ void ap_sendmsg(WORD ap_msg[], WORD type, AESPD *towhom,
     ap_msg[5] = w5;
     ap_msg[6] = w6;
     ap_msg[7] = w7;
-    ap_rdwr(MU_SDMSG, towhom, 16, (LONG)ap_msg);
+    ap_rdwr(MU_SDMSG, towhom, 16, ap_msg);
 }
